refactor: export tier refs functions
