+++ pkg/test/tier/templateRefs.go
@@ -1,0 +1,40 @@
+package tier
+
+import (
+	"crypto/md5"
+	"encoding/hex"
+	"encoding/json"
+	"sort"
+
+	toolchainv1alpha1 "github.com/codeready-toolchain/api/api/v1alpha1"
+)
+
+// ComputeTemplateRefsHash computes the hash of the `.spec.namespaces[].templateRef` + `.spec.clusteResource.TemplateRef`
+func ComputeTemplateRefsHash(tier *toolchainv1alpha1.NSTemplateTier) (string, error) {
+	var refs []string
+	for _, ns := range tier.Spec.Namespaces {
+		refs = append(refs, ns.TemplateRef)
+	}
+	if tier.Spec.ClusterResources != nil {
+		refs = append(refs, tier.Spec.ClusterResources.TemplateRef)
+	}
+	sort.Strings(refs)
+	m, err := json.Marshal(templateRefs{Refs: refs})
+	if err != nil {
+		return "", err
+	}
+	md5hash := md5.New()
+	// Ignore the error, as this implementation cannot return one
+	_, _ = md5hash.Write(m)
+	hash := hex.EncodeToString(md5hash.Sum(nil))
+	return hash, nil
+}
+
+// TemplateTierHashLabel returns the label key to specify the version of the templates of the given tier
+func TemplateTierHashLabelKey(tierName string) string {
+	return toolchainv1alpha1.LabelKeyPrefix + tierName + "-tier-hash"
+}
+
+type templateRefs struct {
+	Refs []string `json:"refs"`
+}
